+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Core updates:
 * General Statistics table now hidden if empty.
 * Fixed some bugs with config settings (eg. template) being overwritten.
 * NFS log file deletion bug fixed by @brainstorm (#265)
+* Fixed bug in `--ignore` behaviour with directory names.
 
 Many thanks to those at the [OpenBio Codefest 2016](https://www.open-bio.org/wiki/Codefest_2016)
 who worked on MultiQC projects.